+++ rust/src/flowgraph/edge.rs
@@ -1,8 +1,8 @@
-use binaryninjacore_sys::*;
-
+use crate::architecture::BranchType;
 use crate::flowgraph::node::FlowGraphNode;
-use crate::flowgraph::{BranchType, EdgePenStyle, ThemeColor};
+use crate::flowgraph::{EdgePenStyle, ThemeColor};
 use crate::rc::{CoreArrayProvider, CoreArrayProviderInner, Ref};
+use binaryninjacore_sys::*;
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct FlowGraphEdge {
